@@ -4,10 +4,11 @@ import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.os.Build;
-import android.support.v7.preference.EditTextPreference;
 import android.util.AttributeSet;
 
 import com.github.tianma8023.xposed.smscode.R;
+
+import androidx.preference.EditTextPreference;
 
 public class ResetEditPreference extends EditTextPreference {
 
