@@ -2,10 +2,11 @@ package com.tianma.xsmscode.common.preference;
 
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.preference.PreferenceDialogFragmentCompat;
 import android.view.View;
 import android.widget.EditText;
+
+import androidx.annotation.NonNull;
+import androidx.preference.PreferenceDialogFragmentCompat;
 
 public class ResetEditPreferenceDialogFragCompat extends PreferenceDialogFragmentCompat {
 
