@@ -7,12 +7,13 @@ import android.content.UriMatcher;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
 import com.github.tianma8023.xposed.smscode.BuildConfig;
 import com.tianma.xsmscode.data.db.entity.SmsCodeRuleDao;
 import com.tianma.xsmscode.data.db.entity.SmsMsgDao;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public class DBProvider extends ContentProvider {
 
