@@ -1,11 +1,12 @@
 package com.tianma.xsmscode.ui.app.base;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 
-import com.tianma.xsmscode.ui.theme.ThemeItemContainer;
 import com.tianma.xsmscode.common.utils.SPUtils;
+import com.tianma.xsmscode.ui.theme.ThemeItemContainer;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 public abstract class BaseActivity extends AppCompatActivity {
 
