@@ -1,8 +1,6 @@
 package com.tianma.xsmscode.ui.faq;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,6 +9,9 @@ import android.widget.TextView;
 import com.github.tianma8023.xposed.smscode.R;
 
 import java.util.List;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class FaqItemAdapter extends RecyclerView.Adapter<FaqItemAdapter.ViewHolder> {
 
