@@ -1,9 +1,6 @@
 package com.tianma.xsmscode.ui.theme;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,6 +10,10 @@ import com.github.tianma8023.xposed.smscode.R;
 import com.tianma.xsmscode.common.adapter.ItemCallback;
 
 import java.util.List;
+
+import androidx.annotation.NonNull;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class ThemeItemAdapter extends RecyclerView.Adapter<ThemeItemAdapter.VH> {
 
